# Push down DELETE for enforceable filters on Delta Lake