<!-- Thank you for submitting a pull request! Find more information in our development guide at https://github.com/trinodb/trino/blob/master/.github/DEVELOPMENT.md and contact us on #dev in Slack. -->
<!-- Provide an overview of the PR for maintainers and reviewers. -->
## Description

Perform DELETE only on the metadata of the Delta Lake
table when the delete filter does not touch any data
columns.

Sample queries affected by this enhancement are:

```
DELETE FROM table 
DELETE FROM table WHERE true
DELETE FROM partitioned_table WHERE part1=value1 AND part2=value2
DELETE FROM partitioned_table WHERE part2=value2
```

Fixes #18331



The operation will not be performed on the metadata layer when
there are `add` file entries in the transaction log of the table
without statistics containing number of records.

<!-- Provide details that would help an engineer who is unfamiliar with this part of the code. -->
## Additional context and related issues


Without `DELETE` pushdown

```
trino:default> create table delta.tiny.sforders100 as select * from tpch.sf100.orders;
CREATE TABLE: 150000000 rows

trino:default> delete from delta.tiny.sforders100 where true;
DELETE: 150000000 rows

Query 20230725_125258_00003_xfv9h, FINISHED, 1 node
http://localhost:8080/ui/query.html?20230725_125258_00003_xfv9h
Splits: 25 total, 25 done (100.00%)
CPU Time: 115.1s total,  1.3M rows/s, 11.2MB/s, 89% active
Per Node: 1.1 parallelism, 1.46M rows/s, 12.5MB/s
Parallelism: 1.1
Peak Memory: 87.3MB
1:43 [150M rows, 1.26GB] [1.46M rows/s, 12.5MB/s]
```

With `DELETE` pushdown

```
trino:default> delete from delta.tiny.sf100orders where true;
DELETE: 150000000 rows

Query 20230725_123039_00000_fk4kt, FINISHED, 1 node
http://localhost:8080/ui/query.html?20230725_123039_00000_fk4kt
Splits: 1 total, 1 done (100.00%)
CPU Time: 0.1s total,     0 rows/s,     0B/s, 83% active
Per Node: 0.1 parallelism,     0 rows/s,     0B/s
Parallelism: 0.1
Peak Memory: 112B
1.84 [0 rows, 0B] [0 rows/s, 0B/s]
```

tldr; before 103 s after 1.8 s


<!-- Mark the appropriate option with an (x). Propose a release note if you can. -->
## Release notes

(x) Release notes are required, with the following suggested text:

```markdown
# Delta Lake
* Improve performance of `DELETE` statement when it deletes the whole table or the filters apply only to the partition columns. ({issue}`18332 `)
```