@findinpath 2023-07-19
`TestDeltaLakeDelete.testDeleteAllOssDeltaLake` is failing because the number of deleted rows is not anymore returned by the `DELETE` query.

Should the accelerated DELETE (done only based on table metadata) be performed if and only if all the data files have the number of records present in the transaction log file?